Conversion to double precision

+++ Source/Application/Constants/Constants.cpp
@@ -1,3 +1,3 @@
 #include "Constants.h"
 
-float TIME_SCALE = 1.0f;
+float64 TIME_SCALE = 1.0;
