+++ Source/Application/Resource/Components/Transform/Position.h
@@ -11,11 +11,11 @@ class Position {
 public:
 	Position()
 	{
-		m_world = Math::Vec3f();
-		m_normalized = Math::Vec3f();
+		m_world = Math::Vec3d();
+		m_normalized = Math::Vec3d();
 	}
 
-	Position(Math::Vec3f position, bool normal = false)
+	Position(Math::Vec3d position, bool normal = false)
 	{
 		normal ? SetNormal(position) : SetWorld(position);
 	}
@@ -23,11 +23,11 @@ public:
 	~Position() = default;
 
 	// Getters
-	const Math::Vec3f& GetWorld() const { return m_world; }
-	const Math::Vec3f& GetNormal() const { return m_normalized; }
+	const Math::Vec3d& GetWorld() const { return m_world; }
+	const Math::Vec3d& GetNormal() const { return m_normalized; }
 
 	// Setters
-	void SetWorld(const Math::Vec3f& position)
+	void SetWorld(const Math::Vec3d& position)
 	{
 		m_world = position;
 
@@ -36,7 +36,7 @@ public:
 		m_normalized.z = position.z / METERS_PER_UNIT;
 	}
 
-	void SetNormal(const Math::Vec3f& position)
+	void SetNormal(const Math::Vec3d& position)
 	{
 		m_world.x = position.x * METERS_PER_UNIT;
 		m_world.y = position.y * METERS_PER_UNIT;
@@ -55,47 +55,47 @@ public:
 	*/
 
 	// Scalar multiplication
-	Position operator*(float scalar) const {
+	Position operator*(double scalar) const {
 		return Position(this->m_world * scalar);
 	}
 
 	// Scalar division
-	Position operator/(float scalar) const {
+	Position operator/(double scalar) const {
 		return Position(this->m_world / scalar);
 	}
 
 	// Scalar addition
-	Position operator+(float scalar) const {
-		return Position(this->m_world + Math::Vec3f(scalar));
+	Position operator+(double scalar) const {
+		return Position(this->m_world + Math::Vec3d(scalar));
 	}
 
 	// Scalar subtraction
-	Position operator-(float scalar) const {
-		return Position(this->m_world - Math::Vec3f(scalar));
+	Position operator-(double scalar) const {
+		return Position(this->m_world - Math::Vec3d(scalar));
 	}
 
 	// In-place versions
-	Position& operator*=(float scalar) {
+	Position& operator*=(double scalar) {
 		this->SetWorld(this->m_world * scalar);
 		return *this;
 	}
 
-	Position& operator/=(float scalar) {
+	Position& operator/=(double scalar) {
 		this->SetWorld(this->m_world / scalar);
 		return *this;
 	}
 
-	Position& operator+=(float scalar) {
-		this->SetWorld(this->m_world + Math::Vec3f(scalar));
+	Position& operator+=(double scalar) {
+		this->SetWorld(this->m_world + Math::Vec3d(scalar));
 		return *this;
 	}
 
-	Position& operator-=(float scalar) {
-		this->SetWorld(this->m_world - Math::Vec3f(scalar));
+	Position& operator-=(double scalar) {
+		this->SetWorld(this->m_world - Math::Vec3d(scalar));
 		return *this;
 	}
 
 private:
-	Math::Vec3f m_world;
-	Math::Vec3f m_normalized;
+	Math::Vec3d m_world;
+	Math::Vec3d m_normalized;
 };
